Add files via upload
This game, generated on the Scratch Program is a game available to all ages and demographics.
The aim of the game is to keep the ball safe from the crabs at the bottom of the screen!
To play the game on a computer device, press the 'space bar' to start, the ball should begin moving with the comment "Use the paddle to save the ball from the crabs!".
To move the red paddle side to side, use the left and right arrow keys in the direction you would like the paddle to move, this is so the ball can bounce off of the paddle and away from the crabs.
The score will be displayed in the top left corner of the screen and will increase every time the ball hits the paddle.
If the ball touches the crabs, a 'pop' noise will occur and a message will appear that reads "You're out! Press 'Space' to restart".
To restart the game, press the space bar and repeat these steps to help keep the ball safe from the crabs!